Modifiche

+++ kernel/resnet50_8/inc/avg_pool.h
@@ -1,9 +1,8 @@
-
 #ifndef AVGPOOL_H
 #define AVGPOOL_H
 
 #include <stdint.h>
 
-void avgpool2x2(const int32_t *in, int H, int W, int32_t *out);
+void avgpool2x2(const int8_t *in, int H, int W, int8_t *out);
 
 #endif
